Refactor Mattermost and Slack authentication providers to use a shared OAuthProvider base class
- Updated MattermostAuthenticationProvider to inherit from OAuthProvider, simplifying the implementation of OAuth flows.
- Introduced OAuthProvider class to encapsulate common OAuth functionality, including token exchange and user fetching.
- Refactored SlackAuthenticationProvider to also inherit from OAuthProvider, ensuring consistency in handling OAuth logic.
- Enhanced error handling and response validation in both providers for improved reliability.
